requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
discord==2.1.0
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
ImageHash==4.3.1
multidict==6.0.2
numpy==1.23.4
Pillow==9.3.0
pycodestyle==2.9.1
python-dotenv==0.21.0
PyWavelets==1.4.1
requests==2.28.1
scipy==1.9.3
sql==2022.4.0
tomli==2.0.1
urllib3==1.26.12
yarl==1.8.1
